+++ beekeeper-path-scheduler-apiary/src/main/java/com/expediagroup/beekeeper/scheduler/apiary/context/CommonBeans.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2019 Expedia, Inc.
+ * Copyright (C) 2019-2020 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,6 +15,7 @@
  */
 package com.expediagroup.beekeeper.scheduler.apiary.context;
 
+import java.util.EnumMap;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -29,16 +30,12 @@ import org.springframework.retry.annotation.EnableRetry;
 import com.expedia.apiary.extensions.receiver.common.messaging.MessageReader;
 import com.expedia.apiary.extensions.receiver.sqs.messaging.SqsMessageReader;
 
-import com.expediagroup.beekeeper.scheduler.apiary.filter.EventTypeListenerEventFilter;
-import com.expediagroup.beekeeper.scheduler.apiary.filter.ListenerEventFilter;
-import com.expediagroup.beekeeper.scheduler.apiary.filter.MetadataOnlyListenerEventFilter;
-import com.expediagroup.beekeeper.scheduler.apiary.filter.TableParameterListenerEventFilter;
-import com.expediagroup.beekeeper.scheduler.apiary.filter.WhitelistedListenerEventFilter;
-import com.expediagroup.beekeeper.scheduler.apiary.messaging.FilteringMessageReader;
-import com.expediagroup.beekeeper.scheduler.apiary.messaging.MessageEventToPathEventMapper;
+import com.expediagroup.beekeeper.core.model.LifecycleEventType;
+import com.expediagroup.beekeeper.scheduler.apiary.handler.MessageEventHandler;
+import com.expediagroup.beekeeper.scheduler.apiary.messaging.BeekeeperEventReader;
 import com.expediagroup.beekeeper.scheduler.apiary.messaging.MessageReaderAdapter;
-import com.expediagroup.beekeeper.scheduler.apiary.messaging.PathEventReader;
 import com.expediagroup.beekeeper.scheduler.apiary.messaging.RetryingMessageReader;
+import com.expediagroup.beekeeper.scheduler.service.SchedulerService;
 
 @Configuration
 @ComponentScan(basePackages = { "com.expediagroup.beekeeper.core", "com.expediagroup.beekeeper.scheduler" })
@@ -47,32 +44,28 @@ import com.expediagroup.beekeeper.scheduler.apiary.messaging.RetryingMessageRead
 @EnableRetry(proxyTargetClass = true)
 public class CommonBeans {
 
-  @Value("${properties.apiary.queue-url}")
-  private String queueUrl;
+  @Bean
+  public EnumMap<LifecycleEventType, SchedulerService> schedulerServiceMap(List<SchedulerService> schedulerServices) {
+    EnumMap<LifecycleEventType, SchedulerService> schedulerMap = new EnumMap<>(LifecycleEventType.class);
+    schedulerServices.forEach(scheduler -> schedulerMap.put(scheduler.getLifecycleEventType(), scheduler));
+    return schedulerMap;
+  }
 
   @Bean(name = "sqsMessageReader")
-  MessageReader messageReader() {
+  public MessageReader messageReader(@Value("${properties.apiary.queue-url}") String queueUrl) {
     return new SqsMessageReader.Builder(queueUrl).build();
   }
 
   @Bean(name = "retryingMessageReader")
-  MessageReader retryingMessageReader(@Qualifier("sqsMessageReader") MessageReader messageReader) {
+  public MessageReader retryingMessageReader(@Qualifier("sqsMessageReader") MessageReader messageReader) {
     return new RetryingMessageReader(messageReader);
   }
 
-  @Bean(name = "filteringMessageReader")
-  MessageReader filteringMessageReader(@Qualifier("retryingMessageReader") MessageReader messageReader,
-    TableParameterListenerEventFilter tableParameterFilter, EventTypeListenerEventFilter eventTypeFilter,
-    MetadataOnlyListenerEventFilter metadataOnlyListenerEventFilter,
-    WhitelistedListenerEventFilter whitelistedListenerEventFilter) {
-    List<ListenerEventFilter> filters = List.of(eventTypeFilter, tableParameterFilter,
-      metadataOnlyListenerEventFilter, whitelistedListenerEventFilter);
-    return new FilteringMessageReader(messageReader, filters);
-  }
-
   @Bean
-  PathEventReader pathEventReader(@Qualifier("filteringMessageReader") MessageReader messageReader,
-    MessageEventToPathEventMapper mapper) {
-    return new MessageReaderAdapter(messageReader, mapper);
+  public BeekeeperEventReader pathEventReader(
+      @Qualifier("retryingMessageReader") MessageReader messageReader,
+      List<MessageEventHandler> handlers
+  ) {
+    return new MessageReaderAdapter(messageReader, handlers);
   }
 }